+++ src/Service/Iam/CHANGELOG.md
@@ -2,13 +2,20 @@
 
 ## NOT RELEASED
 
+## 1.4.0
+
+### Added
+
+- Added `IamClient::listServiceSpecificCredentials()`
+- Added `IamClient::deleteServiceSpecificCredential()`
+
 ## 1.3.0
 
 ### Added
 
 - Added `IamClient::putUserPolicy()`
 - Added `IamClient::deleteUserPolicy()`
-- Added `IamClient::::createServiceSpecificCredential()`
+- Added `IamClient::createServiceSpecificCredential()`
 
 ## 1.2.0